// src/app/models/employee.ts
export interface Employee {
    identificationType: string;
    identificationNumber: string;
    names: string;
    email: string;
    birthdate?: string;
    address?: string;
    city: string;
    phoneNumber?: string;
    birthCountry?: string;
    cellphoneNumber?: string;
    gender?: string;
    position?: string;
    surnames?: string;
    role: string;
    profilePicture?: string;
    workspace?: string;
    company: string;
  }